make changes at user route in backend and implemented regex for email validation in frontend
