--- Backend/.env
@@ -1,2 +1,0 @@
-MONGO_URL=mongodb://127.0.0.1:27017/smartbuy
-PORT=8000
