+++ Backend/README.md
@@ -1,0 +1,11 @@
+# SmartBuy_API
+
+google login =>  deployed_url/user/auth/google, method:get <br>
+Normal login => deployed_url/user/login, method:post <br>
+Signup => deployed_url/user/signup , Method:post <br>
+forget Password => deployed_rul/user/forgotPassword , Method:patch <br>
+logout => deployed/user/logout, Method:delete <br>
+
+product => deployed/product/  <br>
+product add => deployed/product/add
+
