--- Backend/Routes/user_route.js
@@ -1,6 +1,0 @@
-const {registerNewUser}=require("../controllers/user_controller")
-const userRouter=require("express").Router();
-
-userRouter.post("/res",registerNewUser);
-
-module.exports={userRouter}
