--- Backend/config/db.js
@@ -1,10 +1,0 @@
-
-
-let mongoose = require("mongoose")
-require('dotenv').config()
-
-let connection = mongoose.connect(process.env.MONGO_URL)
-
-
-module.exports = {connection}
-
