+++ Backend/db.js
@@ -1,0 +1,7 @@
+let mongoose = require("mongoose")
+require('dotenv').config()
+
+let connection = mongoose.connect(process.env.mongodb)
+
+
+module.exports = {connection}
