+++ Backend/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "backend",
+  "name": "project",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "nodemon main.js"
+    "start": "nodemon server.js"
   },
   "keywords": [],
   "author": "",
@@ -15,8 +15,12 @@
     "cors": "^2.8.5",
     "dotenv": "^16.1.4",
     "express": "^4.18.2",
+    "ioredis": "^5.3.2",
     "jsonwebtoken": "^9.0.0",
-    "mongoose": "^7.2.4",
-    "nodemon": "^2.0.22"
+    "mongoose": "^7.3.0",
+    "nodemailer": "^6.9.3",
+    "nodemon": "^2.0.22",
+    "passport": "^0.6.0",
+    "passport-google-oauth2": "^0.2.0"
   }
 }
