--- Backend_API/.env
@@ -1,6 +1,0 @@
-port = 400
-mongodb = mongodb://localhost:27017/NXM201
-clientID = 52083381668-dteb0gmo91hl42ajn99jg1mce86n21pf.apps.googleusercontent.com
-clientSecret = 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 
-password = dtaecorslfbtvqoq
-privateKey = Authentication
