--- Backend_API/.gitignore
@@ -1,1 +1,0 @@
-node_modules
