--- Backend_API/db.js
@@ -1,7 +1,0 @@
-let mongoose = require("mongoose")
-require('dotenv').config()
-
-let connection = mongoose.connect(process.env.mongodb)
-
-
-module.exports = {connection}
