--- Backend_API/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "project",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcrypt": "^5.1.0",
-    "dotenv": "^16.1.4",
-    "express": "^4.18.2",
-    "ioredis": "^5.3.2",
-    "jsonwebtoken": "^9.0.0",
-    "mongoose": "^7.3.0",
-    "nodemailer": "^6.9.3",
-    "nodemon": "^2.0.22",
-    "passport": "^0.6.0",
-    "passport-google-oauth2": "^0.2.0"
-  }
-}
