--- Backend_API/redis.db.js
@@ -1,7 +1,0 @@
-let Redis = require('ioredis')
-
-let redis = new Redis()
-
-console.log("redis")
-
-module.exports = {redis}
