+++ Frontend/html/Signup.html
@@ -239,11 +239,11 @@
             <h2>Create Account</h2>
             <form>
                 <!-- <label for="Your name">Your name</label><br> -->
-                <input type="text" id="name" class="input" placeholder="Enter Your Name"><br>
+                <input type="text" id="name" class="input" placeholder="Enter Your Name" required><br>
                 <!-- <label for="Email">Email</label><br> -->
-                <input type="Email" id="Email" class="input" placeholder="Enter your e-mail"><br>
+                <input type="email" id="Email" class="input" placeholder="Enter your e-mail" required><br>
                 <!-- <label for="Password">Password</label><br> -->
-                <input type="Password" id="Password" class="input" placeholder="Enter your Password"><br>
+                <input type="password" id="Password" class="input" placeholder="Enter your Password" required><br>
                 <p class="terms"><input type="checkbox" required> By Clicking, you agree to <br> SmartBuy’s Conditions
                     of Use and Privacy Policy.</p>
                 <button id="Signupbtn">Sign Up</button>
@@ -264,9 +264,9 @@
 </script>
 
 <script>
-    let cartData = JSON.parse(localStorage.getItem("cart-data"));
+    let cartData = JSON.parse(localStorage.getItem("cart-data")) || [];
 
-    const cartTotal = document.getElementById("cartTotal");
+    let cartTotal = document.getElementById("cartTotal");
     cartTotal.innerText = cartData.length;
 </script>
