+++ Frontend/index.html
@@ -186,7 +186,7 @@
                         <a href="#" class="nav-link">Customer Service</a>
                     </li>
                     <li class="nav-item">
-                        <a href="#" class="nav-link">Sign Out</a>
+                        <button id="signOutLink" class="nav-link">Sign Out</button>
                     </li>
                     <hr>
                     <div id="menu-bottom">
